Specify how html and body elements fill the viewport
WebKit, Chromium and Edge implement this quirk. The spec was
previously using Gecko's approach, but now Gecko wants to copy
this quirk [1]. I based the spec on Chromium's implementation and
tested in Chromium and Edge.

The spec is stricter than browsers in that it requires the
document element to be an html element, and it requires the
body element to be "the document's body element" (as opposed to
any body), to be consistent with e.g. 'background-color' on body.

Chromium and Edge are inconsistent in their behavior when html
and body have different writing modes:
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4149

Other tests:
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4148
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4146
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4142
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4150
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4139
http://software.hixie.ch/utilities/js/live-dom-viewer/saved/4138

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1259924#c9